Add ShareList components
