+++ src/components/Share/index.js
@@ -1,0 +1,2 @@
+export { default as NoShares } from './NoShares'
+export { default as ShareListItem } from './ShareListItem'
